requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
billiard==4.2.1
boto3==1.37.2
botocore==1.37.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
deprecated==1.2.18
distro==1.9.0
fastapi==0.109.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
kombu==5.4.2
limits==4.0.1
markupsafe==3.0.2
multidict==6.1.0
openai==1.12.0
packaging==24.2
pillow==11.1.0
prompt-toolkit==3.0.50
propcache==0.3.0
pycparser==2.22
pydantic==2.6.1
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.2
redis==5.2.1
replicate==0.22.0
requests==2.31.0
s3transfer==0.11.3
setuptools==75.8.2
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
wsl==0.12
yarl==1.18.3
zope-event==5.0
zope-interface==7.2
logger